fix broker tests separate resolver tests
